split engine module into multiple files and provide a facade
